+++ src/components/script/main.js
@@ -1,3 +1,3 @@
 // import custom components to register them
 import "./navbar";
-import "./modal";
+import "./help-modal";
